Update setup.sh to check if dir exists before removing

+++ setup.sh
@@ -4,4 +4,5 @@
 /usr/bin/python3 -m venv .venv
 . ./.venv/bin/activate
 /usr/bin/python3 -m pip install .
-rm -r ".\r2northstart_moddingdocs.egg-info"
+if test -d ".\r2northstart_moddingdocs.egg-info"; then rm -r ".\r2northstart_moddingdocs.egg-info"
+
